fix: fix numbers display (int in java uses only 32 bits) and remove ugly conversion

+++ src/main/java/jua/objects/LuaNumber.java
@@ -44,10 +44,10 @@ public class LuaNumber implements LuaObject {
   @Override
   public String repr() {
     if (value == Math.floor(value)) {
-      return String.format("%d", value.intValue());
+      return String.format("%d", value.longValue());
     }
 
-    return ((Float) ((float) ((double) value))).toString();
+    return value.toString();
   }
 
   @Override
